+++ tests/unit/ActionListContextProviderTest.tsx
@@ -1,20 +1,22 @@
 import {act, renderHook} from '@testing-library/react-native';
-import useActionListContextValue from '@hooks/useActionListContextValue';
+import ActionListContextProvider from '@hooks/ActionListContextProvider';
+import {useActionListContext} from '@pages/inbox/ReportScreenContext';
 import type {FlatListRefType} from '@pages/inbox/ReportScreenContext';
 
 /**
- * `useActionListContextValue` owns a private holder and exposes `registerListRef(ref)` / `getListRef()`
+ * `ActionListContextProvider` owns a private holder and exposes `registerListRef(ref)` / `getListRef()`
  * plus the scroll refs so each list owns its ref locally while handlers resolve it at call time.
+ * The tests exercise the public surface (the provider + the `useActionListContext` consumer hook).
  */
-describe('useActionListContextValue', () => {
+describe('ActionListContextProvider', () => {
     it('getListRef() returns null before any ref is registered', () => {
-        const {result} = renderHook(() => useActionListContextValue());
+        const {result} = renderHook(() => useActionListContext(), {wrapper: ActionListContextProvider});
 
         expect(result.current.getListRef()).toBeNull();
     });
 
     it('getListRef() returns the same ref object after registerListRef(ref)', () => {
-        const {result} = renderHook(() => useActionListContextValue());
+        const {result} = renderHook(() => useActionListContext(), {wrapper: ActionListContextProvider});
         const listRef: FlatListRefType = {current: null};
 
         act(() => {
@@ -25,7 +27,7 @@ describe('useActionListContextValue', () => {
     });
 
     it('getListRef() returns null again after registerListRef(null) (unmount path)', () => {
-        const {result} = renderHook(() => useActionListContextValue());
+        const {result} = renderHook(() => useActionListContext(), {wrapper: ActionListContextProvider});
         const listRef: FlatListRefType = {current: null};
 
         act(() => {
@@ -41,7 +43,7 @@ describe('useActionListContextValue', () => {
     });
 
     it('keeps scrollPositionRef / scrollOffsetRef stable across re-renders with their initial values', () => {
-        const {result, rerender} = renderHook(() => useActionListContextValue());
+        const {result, rerender} = renderHook(() => useActionListContext(), {wrapper: ActionListContextProvider});
 
         const firstScrollPositionRef = result.current.scrollPositionRef;
         const firstScrollOffsetRef = result.current.scrollOffsetRef;
